chore: reorganise E2E tests (#24)
