+++ SourceFormats/E2E/SourceFormatNode/SourceFormatNodeSdk_Should.cs
@@ -1,0 +1,10 @@
+namespace EncyclopediaGalactica.SourceFormats.E2E;
+
+using System.Diagnostics.CodeAnalysis;
+using Xunit;
+
+[ExcludeFromCodeCoverage]
+[Collection("E2E")]
+public partial class SourceFormatNodeSdk_Should : TestBase
+{
+}
